Moved C files for X11 provider into an x11 package to prevent C file inclusion when not needed, preventing the need for CGO where it isn't needed.
